+++ crates/web-sys/src/features/gen_GpuPipelineLayoutDescriptor.rs
@@ -43,7 +43,7 @@ extern "C" {
     #[wasm_bindgen(method, getter = "bindGroupLayouts")]
     pub fn get_bind_group_layouts(
         this: &GpuPipelineLayoutDescriptor,
-    ) -> ::js_sys::Array<::js_sys::JsOption<GpuBindGroupLayout>>;
+    ) -> ::js_sys::Array<::js_sys::JsNullable<GpuBindGroupLayout>>;
     #[cfg(web_sys_unstable_apis)]
     #[cfg(feature = "GpuBindGroupLayout")]
     #[doc = "Change the `bindGroupLayouts` field of this object."]
@@ -55,7 +55,7 @@ extern "C" {
     #[wasm_bindgen(method, setter = "bindGroupLayouts")]
     pub fn set_bind_group_layouts(
         this: &GpuPipelineLayoutDescriptor,
-        val: &[::js_sys::JsOption<GpuBindGroupLayout>],
+        val: &[::js_sys::JsNullable<GpuBindGroupLayout>],
     );
     #[cfg(web_sys_unstable_apis)]
     #[doc = "Get the `immediateSize` field of this object."]
@@ -85,7 +85,7 @@ impl GpuPipelineLayoutDescriptor {
     #[doc = ""]
     #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
     #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
-    pub fn new(bind_group_layouts: &[::js_sys::JsOption<GpuBindGroupLayout>]) -> Self {
+    pub fn new(bind_group_layouts: &[::js_sys::JsNullable<GpuBindGroupLayout>]) -> Self {
         #[allow(unused_mut)]
         let mut ret: Self = ::wasm_bindgen::JsCast::unchecked_into(::js_sys::Object::new());
         ret.set_bind_group_layouts(bind_group_layouts);
@@ -102,7 +102,7 @@ impl GpuPipelineLayoutDescriptor {
     #[deprecated = "Use `set_bind_group_layouts()` instead."]
     pub fn bind_group_layouts(
         &mut self,
-        val: &[::js_sys::JsOption<GpuBindGroupLayout>],
+        val: &[::js_sys::JsNullable<GpuBindGroupLayout>],
     ) -> &mut Self {
         self.set_bind_group_layouts(val);
         self